altered package structure and pom.xml
