--- inventory/src/main/java/nl/bep3/teamtwee/inventory/InventoryServiceApplication.java
@@ -1,13 +1,0 @@
-package nl.bep3.teamtwee.inventory;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InventoryServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(InventoryServiceApplication.class, args);
-    }
-
-}
