--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/application/command/DeleteIngredient.java
@@ -1,16 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.application.command;
-
-import lombok.Getter;
-
-import java.util.UUID;
-
-@Getter
-public class DeleteIngredient {
-
-    private final UUID id;
-
-    public DeleteIngredient(UUID id) {
-        this.id = id;
-    }
-
-}
