--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/application/command/ReserveIngredients.java
@@ -1,7 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.application.command;
-
-public class ReserveIngredients {
-
-    public ReserveIngredients() {
-    }
-}
