--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/application/query/ListIngredients.java
@@ -1,19 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.application.query;
-
-import lombok.Getter;
-
-@Getter
-public class ListIngredients {
-
-    private final String orderBy;
-    private final String direction;
-
-    public ListIngredients(String orderBy, String direction) {
-        if (orderBy == null) orderBy = "productName";
-        if (direction == null) direction = "asc";
-
-        this.orderBy = orderBy;
-        this.direction = direction;
-    }
-
-}
