--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/domain/Unit.java
@@ -1,9 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.domain;
-
-public enum Unit {
-
-    GRAMS,
-    MILLILITERS,
-    PIECES
-
-}
