--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/domain/exception/IngredientNotFound.java
@@ -1,8 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.domain.exception;
-
-public class IngredientNotFound extends RuntimeException {
-
-    public IngredientNotFound(String message) {
-        super(message);
-    }
-}
