--- inventory/src/main/java/nl/bep3/teamtwee/inventory/core/domain/exception/InsufficientStockCapacity.java
@@ -1,8 +1,0 @@
-package nl.bep3.teamtwee.inventory.core.domain.exception;
-
-public class InsufficientStockCapacity extends RuntimeException {
-
-    public InsufficientStockCapacity(String message) {
-        super(message);
-    }
-}
