--- inventory/src/main/java/nl/bep3/teamtwee/inventory/infrastructure/driver/web/request/RegisterIngredientRequest.java
@@ -1,27 +1,0 @@
-package nl.bep3.teamtwee.inventory.infrastructure.driver.web.request;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Positive;
-
-public class RegisterIngredientRequest {
-    @NotBlank
-    public String productName;
-
-    @NotBlank
-    public String unit;
-
-    @Positive
-    public int capacity;
-
-    @Positive
-    public int purchaseCapacity;
-
-    @Positive
-    public int sellCapacity;
-
-    @Positive
-    public double purchasePrice;
-
-    @Positive
-    public double sellPrice;
-}
