+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/InventoryServiceApplication.java
@@ -1,0 +1,13 @@
+package nl.bep3.teamtwee.restaurant.inventory;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class InventoryServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(InventoryServiceApplication.class, args);
+    }
+
+}
