+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/application/command/ReserveIngredients.java
@@ -1,0 +1,7 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.application.command;
+
+public class ReserveIngredients {
+
+    public ReserveIngredients() {
+    }
+}
