+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/application/query/FindIngredientsByProductName.java
@@ -1,0 +1,20 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.application.query;
+
+import lombok.Getter;
+
+@Getter
+public class FindIngredientsByProductName {
+
+    private final String productName;
+    private final String orderBy;
+    private final String direction;
+
+    public FindIngredientsByProductName(String productName, String orderBy, String direction) {
+        if (orderBy == null) orderBy = "productName";
+        if (direction == null) direction = "asc";
+        this.productName = productName;
+        this.orderBy = orderBy;
+        this.direction = direction;
+    }
+
+}
