+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/domain/Unit.java
@@ -1,0 +1,9 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.domain;
+
+public enum Unit {
+
+    GRAMS,
+    MILLILITERS,
+    PIECES
+
+}
