+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/domain/exception/IngredientNotFound.java
@@ -1,0 +1,8 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.domain.exception;
+
+public class IngredientNotFound extends RuntimeException {
+
+    public IngredientNotFound(String message) {
+        super(message);
+    }
+}
