+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/domain/exception/InsufficientStockCapacity.java
@@ -1,0 +1,8 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.domain.exception;
+
+public class InsufficientStockCapacity extends RuntimeException {
+
+    public InsufficientStockCapacity(String message) {
+        super(message);
+    }
+}
