+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/core/port/storage/IngredientRepository.java
@@ -1,0 +1,15 @@
+package nl.bep3.teamtwee.restaurant.inventory.core.port.storage;
+
+import nl.bep3.teamtwee.restaurant.inventory.core.domain.Ingredient;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface IngredientRepository extends MongoRepository<Ingredient, UUID> {
+
+    List<Ingredient> findByProductNameEquals(String productName, Sort sort);
+    List<Ingredient> findByProductNameEquals(String productName);
+
+}
