+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/infrastructure/config/WebSecurityConfig.java
@@ -1,0 +1,25 @@
+package nl.bep3.teamtwee.restaurant.inventory.infrastructure.config;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.savedrequest.NullRequestCache;
+import org.springframework.web.cors.CorsConfiguration;
+
+import java.util.Arrays;
+
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE"));
+
+        http.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and().
+                authorizeRequests().anyRequest().permitAll().and().
+                requestCache().requestCache(new NullRequestCache()).and().
+                cors().configurationSource(request -> configuration.applyPermitDefaultValues()).and().
+                csrf().disable();
+    }
+}
