+++ inventory/src/main/java/nl/bep3/teamtwee/restaurant/inventory/infrastructure/driver/web/request/ReserveIngredientRequest.java
@@ -1,0 +1,13 @@
+package nl.bep3.teamtwee.restaurant.inventory.infrastructure.driver.web.request;
+
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.PositiveOrZero;
+import java.util.UUID;
+
+public class ReserveIngredientRequest {
+    @NotNull
+    public UUID id;
+
+    @PositiveOrZero
+    public int amount;
+}
